Deduplicate common memory resource test definition
Right now, we have pretty much exactly the same tests for the core
resources as well as the CUDA ones, the HIP ones, and the SYCL ones.
That's a lot of code duplication! This commit attempts to fix that
problem by lifting all the testing code into one common definition,
which is then used by the different testing executables.

I use header files for this, which contain the test definitions. I then
import them into the test files. Technically there could be some linker
problems with this, but in the tests I feel like this is not at much of
a problem.
