+++ tests/sycl/test_sycl_memory_resources.cpp
@@ -7,7 +7,8 @@
 
 // Local include(s).
 #include "../common/memory_resource_name_gen.hpp"
-#include "vecmem/containers/vector.hpp"
+#include "../common/tests/test_generic_basic.hpp"
+#include "../common/tests/test_generic_host_accessible.hpp"
 #include "vecmem/memory/sycl/device_memory_resource.hpp"
 #include "vecmem/memory/sycl/host_memory_resource.hpp"
 #include "vecmem/memory/sycl/shared_memory_resource.hpp"
@@ -15,36 +16,13 @@
 // GoogleTest include(s).
 #include <gtest/gtest.h>
 
-// System include(s).
-#include <algorithm>
-#include <cmath>
-#include <vector>
-
 // Memory resources.
 static vecmem::sycl::device_memory_resource device_resource;
 static vecmem::sycl::host_memory_resource host_resource;
 static vecmem::sycl::shared_memory_resource shared_resource;
 
-/// Base test case for the SYCL memory resources
-///
-/// This just makes sure that the memory resources defined in the
-/// @c vecmem::sycl library are more-or-less functional.
-///
-class sycl_memory_resource_test
-    : public testing::TestWithParam<vecmem::memory_resource*> {};
-
-/// Perform some very basic tests that do not need host accessibility
-TEST_P(sycl_memory_resource_test, allocations) {
-
-    vecmem::memory_resource* resource = GetParam();
-    for (std::size_t size = 1000; size < 100000; size += 1000) {
-        void* ptr = resource->allocate(size);
-        resource->deallocate(ptr, size);
-    }
-}
-
 // Instantiate the allocation tests on all of the resources.
-INSTANTIATE_TEST_SUITE_P(sycl_memory_resource_tests, sycl_memory_resource_test,
+INSTANTIATE_TEST_SUITE_P(sycl_memory_resource_tests, memory_resource_test_basic,
                          testing::Values(&device_resource, &host_resource,
                                          &shared_resource),
                          vecmem::testing::memory_resource_name_gen(
@@ -52,99 +30,9 @@ INSTANTIATE_TEST_SUITE_P(sycl_memory_resource_tests, sycl_memory_resource_test,
                               {&host_resource, "host_resource"},
                               {&shared_resource, "shared_resource"}}));
 
-namespace {
-
-/// Custom non-trivial type used in the tests.
-struct TestType {
-    TestType(int a, long b = 123) : m_a(a), m_b(b) {}
-    int m_a;
-    long m_b;
-};
-/// Helper operator for @c TestType
-bool operator==(const TestType& value1, const TestType& value2) {
-    return ((value1.m_a == value2.m_a) && (value1.m_b == value2.m_b));
-}
-
-/// Comparison operator for fundamental types
-template <typename T>
-bool almost_equal(const T& value1, const T& value2) {
-    return (std::abs(value1 - value2) < 0.001);
-}
-
-/// Comparison operator for the custom test type
-template <>
-bool almost_equal<TestType>(const TestType& value1, const TestType& value2) {
-    return (value1 == value2);
-}
-
-}  // namespace
-
-/// Test case for host-accessible SYCL memory resources
-///
-/// Providing a slightly more elaborate test for memory resources that can be
-/// read/written from host code.
-///
-class sycl_host_accessible_memory_resource_test
-    : public sycl_memory_resource_test {
-
-protected:
-    /// Function performing some basic tests using @c vecmem::vector
-    template <typename T>
-    void test_host_accessible_resource(vecmem::vector<T>& test_vector) {
-
-        // Set up the test vector, and create a reference vector.
-        std::vector<T> reference_vector;
-        reference_vector.reserve(100);
-        test_vector.reserve(100);
-
-        // Fill them up with some dummy content.
-        for (int i = 0; i < 20; ++i) {
-            reference_vector.push_back(i * 2);
-            test_vector.push_back(i * 2);
-        }
-        // Make sure that they are the same.
-        EXPECT_EQ(reference_vector.size(), test_vector.size());
-        EXPECT_TRUE(std::equal(reference_vector.begin(), reference_vector.end(),
-                               test_vector.begin()));
-
-        // Remove a couple of elements from the vectors.
-        for (int i : {26, 38, 25}) {
-            (void)std::remove(reference_vector.begin(), reference_vector.end(),
-                              i);
-            (void)std::remove(test_vector.begin(), test_vector.end(), i);
-        }
-        // Make sure that they are still the same.
-        EXPECT_EQ(reference_vector.size(), test_vector.size());
-        EXPECT_TRUE(std::equal(reference_vector.begin(), reference_vector.end(),
-                               test_vector.begin(), ::almost_equal<T>));
-    }
-
-};  // class sycl_host_accessible_memory_resource_test
-
-/// Test the host accessible memory resource with an integer type.
-TEST_P(sycl_host_accessible_memory_resource_test, int_value) {
-
-    vecmem::vector<int> test_vector(GetParam());
-    test_host_accessible_resource(test_vector);
-}
-
-/// Test the host accessible memory resource with a floating point type.
-TEST_P(sycl_host_accessible_memory_resource_test, double_value) {
-
-    vecmem::vector<double> test_vector(GetParam());
-    test_host_accessible_resource(test_vector);
-}
-
-/// Test the host accessible memory resource with a custom type.
-TEST_P(sycl_host_accessible_memory_resource_test, custom_value) {
-
-    vecmem::vector< ::TestType> test_vector(GetParam());
-    test_host_accessible_resource(test_vector);
-}
-
 // Instantiate the full test suite on the host-accessible memory resources.
 INSTANTIATE_TEST_SUITE_P(sycl_host_accessible_memory_resource_tests,
-                         sycl_host_accessible_memory_resource_test,
+                         memory_resource_test_host_accessible,
                          testing::Values(&host_resource, &shared_resource),
                          vecmem::testing::memory_resource_name_gen(
                              {{&host_resource, "host_resource"},
